Move CPU query into separate file
