+++ package/resource_monitor/src/Makefile
@@ -1,5 +1,5 @@
 TARGET=resource_monitor
-SOURCES=resource_monitor.c resource_query.c
+SOURCES=resource_monitor.c resource_query.c query_cpu_usage.c
 LIBS=-luv -lnanomsg -lsbp -lpiksi
 CFLAGS=-std=gnu11 -Wmissing-prototypes -Wconversion -Wimplicit -Wshadow -Wswitch-default -Wswitch-enum -Wundef -Wuninitialized -Wpointer-arith -Wstrict-prototypes -Wcast-align -Wformat=2 -Wimplicit-function-declaration -Wredundant-decls -Wformat-security -Wall -Wextra -Wno-strict-prototypes -Werror
